requirements.txt
decorator==4.0.6
Django==1.9.1
django-extensions==1.6.1
django-filter==0.12.0
django-jsonfield==0.9.19
djangorestframework==3.3.2
ipython==4.0.3
ipython-genutils==0.1.0
Markdown==2.6.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
ply==3.8
psycopg2==2.6.1
ptyprocess==0.5
Pykka==1.2.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
Werkzeug==0.11.3
wheel==0.24.0
